refactor(offers): dedupe wrap/filter into shared util, use canonical DTOs
Address PR review on #981:

- Move wrapProductVariantsWithOffers + the offered-products filter into a
  shared packages/core/src/api/utils/offers.ts, reused by both the vendor
  and admin product endpoints. wrap takes an optional sellerId (vendor:
  active seller; admin: none / platform-wide); the filter reads the seller
  scope from req.seller_context (vendor) or the seller_id query param
  (admin Store filter). Remove the admin helpers.ts duplicate and the
  vendor applySellerOfferedProductsFilter copy.
- Extend OfferDTO with the documented-but-missing seller / product_variant
  / shipping_profile relations; admin offer types now compose OfferDTO /
  ProductVariantDTO instead of local shapes.

Build 9/9; offer-products suites (admin + vendor) 13 tests pass.
