+++ packages/core/src/api/utils/index.ts
@@ -5,3 +5,4 @@ export * from "./filter-by-seller-id"
 export * from "./vendor-cors-middleware"
 export * from "./format-product-attributes"
 export * from "./filter-attributes-by-category-link"
+export * from "./offers"
